Fix flags, CMake and build locations
libqi used to expect using gnu++11 C library, and -fPIC.
Let us use it for now for compatibility,
and see if it can be removed later.

The CMake code generated by qitoolchain was not enough for providing
all the details the usual toolchains gives about Boost.
Let us recover the usual toolchains' CMake code to ensure compatibility.
Note that it does not correspond to CMake's usual FindBoost.cmake.

Build locations were modified to reduce the need for `sudo`-ing.
Installation to the system is done after qitoolchain packaging.